Assignment2_5_Lab5_fix_float_to_double

+++ Lab5/CircleLawn.cpp
@@ -25,7 +25,7 @@ namespace lab5
 
 	unsigned int CircleLawn::GetArea() const
 	{
-		float area = mHalfLength * mHalfLength * mPi;
+		double area = mHalfLength * mHalfLength * mPi;
 
 		area += 0.5f;
