+++ Lab5/RectangleLawn.cpp
@@ -33,7 +33,7 @@ namespace lab5
 
 	unsigned int RectangleLawn::GetMinimumFencesCount() const
 	{
-		float fenceCount = ((mHeight + mLength) * 2)  / 0.25f;
+		double fenceCount = ((mHeight + mLength) * 2)  / 0.25f;
 
 		if (fenceCount == static_cast<unsigned int>(fenceCount))
 		{
@@ -46,16 +46,16 @@ namespace lab5
 
 	unsigned int RectangleLawn::GetFencePrice(eFenceType fenceType) const
 	{
-		float fencePrice = 0;
+		double fencePrice = 0;
 
 		switch (fenceType)
 		{
 		case 0:
-			fencePrice = static_cast<float>(GetMinimumFencesCount()) * 1.5f;
+			fencePrice = static_cast<double>(GetMinimumFencesCount()) * 1.5f;
 			break;
 
 		case 1:
-			fencePrice = static_cast<float>(GetMinimumFencesCount()) * 1.75f;
+			fencePrice = static_cast<double>(GetMinimumFencesCount()) * 1.75f;
 			break;
 
 		}
